Avoid realm-level serialization when publishing.
Since realm session handlers need to be able to run concurrently
(especially for publishing), this means that the default broker
and dealer have to be made safe for concurrent use.  To avoid API
changes locks are used at the lower levels rather than goroutines.

In particular a RWMutex is used in the default broker so that
multiple sessions may publish concurrently (read-lock over the maps).
True serialization (write-lock over the maps) is only required during
subscribe/unsubscribe.