Make the EventHandler emit the topic
This way one can implement a generic event handler that can cope with data
coming from an (at compile time) unknown number of topics. A generic event
handler like this might need to know which topic it is dealing with but if the
router doesn't emit this information it is nowhere else available.